backend/requirements.txt
torch
torchvision
uvicorn
fastapi
pillow
pymongo
python-dotenv
